requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0
pillow>=10.0.0

# Data Processing
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.3.2

# Geometry
shapely>=2.0.0

# Database
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0

# Additional dependencies
werkzeug>=3.0.0

